Split build metadata into two scripts.
